# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CryptoCampus API
  version: 1.0.0
  description: |
    团队 01 的 HTTP 接口唯一机器可读契约。所有业务接口以 /api/v1 为前缀。
    修改字段、类型、错误码或权限前，必须先评审本文件。
...
servers:
  - url: /api/v1
    description: 当前部署
...
paths:
...
  /admin/engine:
    get:
      tags: [Admin]
      operationId: getEngineDetails
      summary: 获取引擎与 Provider 详细状态
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      responses:
        "200":
          description: 引擎状态
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemStatus" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/benchmarks:
    post:
      tags: [Admin]
      operationId: runBenchmark
      summary: 创建异步性能基准任务
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [iterations]
              properties:
                iterations: { type: integer, minimum: 10, maximum: 10000, default: 1000 }
                include_pqc: { type: boolean, default: true }
      responses:
        "202":
          description: 已创建任务
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /admin/users:
    get:
      tags: [Admin]
      operationId: listAdminUsers
      summary: 查询用户、角色、状态和允许查看的元信息
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [password, private_key, auth_hash, salt_a, salt_k, refresh_token]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 用户分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{user_id}/status:
    patch:
      tags: [Admin]
      operationId: updateUserStatus
      summary: 冻结或解冻用户并记录审计日志
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [refresh_token, private_key]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status, reason]
              properties:
                status: { type: string, enum: [active, frozen] }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: 更新后的用户
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /admin/users/{user_id}/role:
    patch:
      tags: [Admin]
      operationId: updateUserRole
      summary: 修改角色、吊销旧 Token 并留痕
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [refresh_token]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role, reason]
              properties:
                role: { type: string, enum: [student, admin, teacher] }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: 更新后的用户
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{user_id}/quotas/reset:
    post:
      tags: [Admin]
      operationId: resetUserQuotas
      summary: 重置用户额度并记录审计日志
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 重置后的额度
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuotaPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /admin/hole/posts/{post_id}/withdraw:
    post:
      tags: [Admin]
      operationId: withdrawHolePost
      summary: 撤下违规帖子并追加链式撤销记录
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [author_identity]
      parameters:
        - $ref: "#/components/parameters/PostId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "200":
          description: 撤销记录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RevocationEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /admin/drops/{code}/destroy:
    post:
      tags: [Admin]
      operationId: destroyDrop
      summary: 强制销毁密信密文；管理员不能解密内容
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [ciphertext, plaintext, session_key]
      parameters:
        - $ref: "#/components/parameters/DropCode"
      responses:
        "204": { description: 密文已销毁 }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/votes/{vote_id}/audit-flags:
    post:
      tags: [Admin]
      operationId: flagVoteForAudit
      summary: 标记异常投票供审计，不改变匿名选票归属
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [voter_identity]
      parameters:
        - $ref: "#/components/parameters/VoteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        "201":
          description: 审计标记已记录
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Accepted" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/inspect-records:
    get:
      tags: [Admin]
      operationId: listInspectRecordMetadata
      summary: 查询他人透视记录元信息，不返回敏感中间值
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [plaintext, private_key, kek, session_key, full_intermediate_values]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 仅元信息分页
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InspectMetadataPage" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/providers/reload:
    post:
      tags: [Admin]
      operationId: reloadPqcProvider
      summary: 演示重新加载 PQC Provider
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [provider_configuration_secret]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 重载后的 Provider 状态
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SystemStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /admin/benchmarks/{benchmark_id}:
    get:
      tags: [Admin]
      operationId: getBenchmark
      summary: 获取均值、P99 和抗量子开销对比
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      parameters:
        - $ref: "#/components/parameters/BenchmarkId"
      responses:
        "200":
          description: 基准结果
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BenchmarkResult" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/benchmarks/{benchmark_id}/export:
    get:
      tags: [Admin]
      operationId: exportBenchmark
      summary: 导出已完成的性能基准结果
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      parameters:
        - $ref: "#/components/parameters/BenchmarkId"
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [csv, json] }
      responses:
        "200":
          description: CSV 或 JSON 性能基准结果
          headers:
            Content-Disposition:
              description: 下载文件名
              schema: { type: string }
          content:
            text/csv:
              schema: { type: string }
            application/json:
              schema: { $ref: "#/components/schemas/BenchmarkResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /admin/audit/export:
    get:
      tags: [Admin]
      operationId: exportAdminAudit
      summary: 导出经 SM3 链式哈希保护的管理员审计记录
      x-required-roles: [admin, teacher]
      x-jwt: true
      x-sensitive-fields: [plaintext, private_key, kek, session_key]
      parameters:
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [csv, json] }
      responses:
        "200":
          description: CSV 或 JSON 审计导出
          content:
            text/csv:
              schema: { type: string }
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEntry" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: cc_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    DropCode:
      name: code
      in: path
      required: true
      schema: { type: string, pattern: "^[A-Za-z0-9]{11,32}$" }
...
    UserId:
      name: user_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PostId:
      name: post_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: 未认证、凭据错误或会话失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: 资源不存在或为防枚举隐藏存在性
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 状态冲突、重复消费或幂等冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: 字段校验失败
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: 额度或频率超限
      headers:
        Retry-After:
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServiceUnavailable:
      description: 密码引擎或 Provider 暂不可用
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: 后端或密码引擎内部错误；消息不得包含密钥、明文或原始堆栈
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, role, status, pqc_mode, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { $ref: "#/components/schemas/Email" }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/UserStatus" }
        pqc_mode: { type: boolean }
        created_at: { type: string, format: date-time }
    SystemStatus:
      type: object
      additionalProperties: false
      required: [api, engine, version, tlcp, providers]
      properties:
        api: { type: string, enum: [ok, degraded] }
        engine: { type: string, enum: [online, offline, degraded] }
        version: { type: string }
        tlcp: { type: string, enum: [online, offline, unknown] }
        providers:
          type: object
          additionalProperties: { type: boolean }
...
    QuotaPage:
      type: object
      additionalProperties: false
      required: [items, resets_at]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Quota" }
        resets_at: { type: string, format: date-time }
...
    Job:
      type: object
      additionalProperties: false
      required: [id, status, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, completed, failed] }
        created_at: { type: string, format: date-time }
...
    RevocationEntry:
      type: object
      additionalProperties: false
      required: [sn, reason, hash_prev, hash_curr, timestamp]
      properties:
        sn: { type: string, pattern: "^[A-Fa-f0-9]{32,}$" }
        reason: { type: string, maxLength: 500 }
        hash_prev: { type: string, contentEncoding: base64 }
        hash_curr: { type: string, contentEncoding: base64 }
        timestamp: { type: string, format: date-time }
...
    UserPage:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/User" } }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
security:
  - bearerAuth: []